+++ src/MagicOnion.Server/Binder/Internal/MagicOnionGrpcMethodBinder.cs
@@ -1,8 +1,8 @@
-using System.Reflection;
 using Grpc.AspNetCore.Server.Model;
 using Grpc.Core;
 using MagicOnion.Internal;
 using MagicOnion.Serialization;
+using MagicOnion.Server.Features.Internal;
 using MagicOnion.Server.Hubs.Internal;
 using MagicOnion.Server.Internal;
 using Microsoft.AspNetCore.Routing;
